Removed Tile Actor class. Swapped to InstancedStaticMesh
The Tile Actor class has been removed and instead the tile meshes are handled by the map class which will have a different instanced mesh for each tile type.
Tile type data is now stored in a UDataTable.
The tile map is now stored as a TMap between map coordinates and tile type enums. The properties of a given tile type can then be looked up in the data table.
Some functionality has been removed from the Pawn class while working on this. Eventually this will go back in and I will have to find a way of clicking on an instance of an instanced static mesh and getting its corresponding map coordinate.