Added tile selection and highlighting and some useful map functions
Added a focused tile to the tilemap class, this is effectively the map cursor.
A set of moveable and attackable tiles has been added, with corresponding instanced static meshes. This is for higlighting tiles which can be selected for movement or attacking.
I made some meshes using blender for the tile cursor and a target.
Completely removed the Tile class.
Renamed TileBasedGamePawn to PlayerPawn and updated functionality to new tile selection system.
Unit ACharater class has been made but has nothing in yet.