TileData now TileType. TileType enum removed
TileType enum removed and instead just use the tile id which is its position in the tile data array.
TileData struct renamed TileType as felt more appropriate.
Spent most of today making meshes and materials for the different tile types.
Made materials and meshes for the tile highlighting. These always render on top of the tile meshes despite being lower in the scene