Custom Map Importing
Custom maps can now be imported by painting an image an importing it as a texture. If the color of a pixel matches any of the tile types in the tile types data table then it will add a tile at the matching coordinate.
TileTypes now have a FColor SourceImageColour field to facilitate this.