Unit Basics
Implementation of the AUnit class has begun.
Units have a number of integer stats which can be accessed through Get() functions . Buffs will be implemented which will effect the result returned by the Get function.
The Unit class also has OnTurnStart and OnTurnEnd functions that will enable and disable the unit as well as applying any effects of the buffs.
The ApplyDamage and ApplyHeal functions take a input value of dmg or heal and apply it, capping the health between 0 and max.
The ApplyDamage function also takes a boolean signifying magic damage to determine if the incoming damage should be affected by armour or magic resist.
The TriggerDeath function is called when the units health is reduced to 0 during ApplyDamage. In this function all the necessary cleanup will be performed, death animations ran, and the unit will be destroyed.
The unit will have an array of the abilities that are available to it and the functions CanUseAbility and UseAbility which enable the casting of the abilities.