A* Pathfinding
A* Pathfinding has been implemented. A sequence of map coordinates from a start tile to a target tile is returned by the function GetShortestPath. This takes into account the terrain movement costs and the blocking of units by the enemy team.
To implement this a MoveSequence class was created. This acts like a singly linked list, with some variables for the corresponding tile properties and A* pathfinding values.
The PlayerPawn class is configured for testing this. Clicking on the map will select a starting tile and then the shortest path to wherever the cursor subsequently hovers over is the highlighted.